feat: 航司final

+++ src/main/java/com/yumefusaka/SmartAirport/pojo/DTO/AddFlightDTO.java
@@ -12,7 +12,7 @@ public class AddFlightDTO {
 
     private String flight_number;
 
-    private long capacity;
+    private int capacity;
 
     private String departure_city;
